requirements.txt
defusedxml==0.6.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==2.2.4
django-crispy-forms==1.7.2
django-filter==2.2.0
django-gitrevision==0.1.5
django-simple-history==2.7.3
djangorestframework==3.10.2
gitdb==0.6.4
GitPython==1.0.1
gunicorn==19.3.0
heroku==0.1.4
Markdown==2.6.6
oauthlib==2.0.2
psycopg2==2.8.3
PyJWT==1.5.0
pystache==0.5.4
python-dateutil==1.5
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
smmap==0.9.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
static==1.1.1
static3==0.6.1
